Declaring Mesh, Track, ... Objects
I included a mesh object, this will be used to store QGraphicsItemGroups (e.g. the track, the player animation, clouds, ... ). Furthermore I created a class Track class: Will be used to random generate a FIXED (no movement) track (which will be able to be added to the Mesh object as it had the possibility to output to a QGraphicsItemGroup object containing all items).